test(python): split the API-requirements file by subject
`test_api_requirements.py` carried 58 of the suite's 108 tests under a name
that is history: a Document-structure/`$ext` half and a writer/reader half.
They become `test_document.py` and `test_writer_reader.py`; the engine and
quill tests join `test_quill.py`, and the blank-canvas render flow joins
`test_render.py`.

Four helpers defined two and three times move to `conftest.py`: `field` and
`has_field` (also in `test_parse.py`), `field_keys`, and `_fields_of`
(`test_validate.py`), which is `field` under another name. `test_parse.py`'s
own `QUILLS_PATH` block goes — nothing in the file read it. The `taro` and
`richtext_form` quill loaders, now wanted by three files, move with them.

Eleven tests go, none of them the only statement of what it checks:

- `test_full_workflow`, `test_quill_from_path` and
  `test_insert_card_invalid_kind` restate a neighbour.
- `test_writer_front_door_set_and_reads` restates three reader tests; its
  writer-handle identity assertion lands beside the reader's.
- The four positional card verbs merge into the mutation sequence, which now
  reads the order back after each step rather than only checking that the
  kinds are lowercase, and the two `card=` selector tests become one over
  both lanes.
- `test_schema.py`'s two blueprint absence tests, and the `delete-ok` and
  `field_absent` absence assertions inside two more: core owns the blueprint
  format, and a retired spelling staying retired is regression against
  history.

Closes #1683
